Fix kernel symbol lookup, and thus stack().
dt_symbol_by_addr(), the symbol lookup component of dt_symtab.c, was extremely
buggy.  It did not consider addresses to match a symbol unless they were
identical to the symbol's starting address, rather than considering all
addresses within the range from the starting address to the size to match.
Also, it missed a level of indirection in both the argument to the bsearch()
comparison function and the return value of bsearch() itself, leading to garbage
symbol records.

With this fixed, stack() works as intended: NEWS is updated accordingly.