Save memory by packing symbol names into a string table.
DTrace userspace reads and stores information about all symbols in the running
kernel, including those in modules: many tens of thousands of them, at least.
This uses quite a lot of memory.  We can save much of it by packing the names of
the symbols into a string table, after which we only have one lot of malloc()
overhead per symbol, for the dt_symbol structure, which is of fixed size so
probably dealt with fairly efficiently.

On my system, judicious insertion of malloc_stats() calls suggests that this
saves about 750Kb, a not insignificant part given a total of 5Mb of allocations
at the time.

This does not change the lifetime of symbols returned from dt_symbol_name() or
anywhere else, except that such symbols cease to be valid over the first call to
dt_symtab_pack().  Since DTrace packs symbol tables the first chance it gets,
before returning from the first dtrace_update() call, unpacked symbol names
never make it to the public API and this changes nothing observable by libdtrace
consumers.  (Should we ever start responding to module unloads by discarding the
dt_module state, this might change, but this seems quite unlikely.)