Fixes to Psymtab.c mapfile reading.
The mapfile reading code in libproc's Pupdate_maps() was quite
thoroughly broken.  Not only did it use sscanf() to read in lines from
/proc/$pid/maps, but it used a fixed-size buffer for the permissions
element that was one byte too short, and used a fixed 64-byte buffer
for the filename.  The former is obviously wrong and the latter is a
security hole waiting to happen.  Thankfully, C99 provides a %ms
sscanf() format conversion that dynamically allocates its output, so
closing this hole is trivial.

This hole wasn't ever reached, though, because of a != where a == was
meant, leading to the maps getting reset and thrown away *unless*
opening /proc/self/maps failed, rather than *if* it failed.

(At the same time, clean up some stuff meant for tracking the fake ELF
files dropped in the previous commit: they were entangled with code in
the initialization and cleanup functions we had to change as part of
this commit, and disentangling them seems not worth it.)