Make dtrace -G work.
The drti.c code has been reworked to use /proc/<PID>/maps to determine the
full path of the executable and the base address.  It will now correctly
pass the embedded DOF object to the DTrace helper interface at executable
startup (constructor), and call the remove function at (normal) termination
(destructor).

Small change to dt_link.c to ensure that rewritten USDT probe locations are
not relocated later in the executable link process.  Mark them as type 0
(universal NONE type), which ensures they are ignored.  On OpenSolaris this
was handled through DTrace-aware code in the linker which is entirely
unnecessary.