Support source files in foreign directories.
Before now, if you specified foo/bar.c in name_SOURCES, the build system would
strip the pathame off, and look up only bar.c.  This is counterintuitive at
best, but worse yet, this behaviour applies to name_SRCDEPS, making it
impossible for a source file to depend on a source file (such as a generated
header file) in another directory.

A few places in the build system actually depended on this behaviour, passing
absolute paths into $(src-name) et al and expecting the path to be stripped
off.  This is clearly wrong and is now fixed.

For consistency, so that you don't need to guess whether a given path is
absolute or not, we absolutize $(objdir) as well, and fix the places that
depended on its being relative.