rtld_db support.
This commit introduces an improved implementation of the parts of the Solaris
librtld_db API used by DTrace into libproc, and adjusts DTrace to use it.  It is
improved in that the rtld activity monitor simply calls a callback to indicate a
dlopen()/dlclose(), rather than requiring redundant thrashing about with
breakpoints in every user.  Many fields in rd_loadobj are missing: some have
different interpretations (in particular, rl_base indicates the library's load
bias, not the base address).

It is not a complete implementation of rtld_db: in particular, there is no
support for PLT resolution at all.  Also, the activity monitor will never repot
RD_PREINIT or RD_POSTINIT.  lmid support is present but needs enhancement:
currently, it will only report lmids of zero.

Internally, this is made relatively complex by the need for tracing of 32-bit
processes and the need to consult structures in glibc which are not
public. These are handled by constructing tables of offsets at build time (it is
populated at runtime, but from compile-time computed constant offsets).  Since
this offset is built with a 32-bit compiler, building DTrace now requires a
32-bit glibc and libgcc.

The offsets table consists of a series of #defines giving the offset of various
fields of interest, and a set of arrays, one per structure of interest, mapping
field offsets on the host (64-bit) architecture to offsets and sizes on the
target.  The target offset and size arrays for each source offset are two
elements long: the first element is for 32-bit, the second for 64.  To use it,
the ps_prochandle now contains an 'elf64' field which is 1 when tracing a 64-bit
executable: you can just use this as an index directly.  The arrays are
initialized by a constructor in the machine-generated libproc/rtld_offsets.c.
There is a new Pread_scalar() function which is expected to be used to read a C
primitive type of a particular size into a variable of the appropriate size: it
does appropriate size conversions given the machine endianness.

At rd_new() time, rtld_db looks up the location of _r_debug on the target, using
the DT_DEBUG dynamic tag filled in by ld.so, and if that fails (as with
statically linked binaries, where there is no dynamic section at all), falls
back to looking up the _r_debug symbol directly, using the Pxlookup_by_name
machinery.  (That machinery does not work, as of this commit, but this is a rare
edge case on modern systems).

Givne the location of _r_debug, we can dig out the r_brk breakpoint called by
the dynamic linker whenever a dlopen() or dlclose() is about to happen, or has
finished.  We need this for everything rtld_db does: obviously it is needed for
dlactivity reporting, but it's also needed for link map iteration, because we
cannot iterate over the link maps while ld.so is modifying them.

Digging out _r_debug and r_brk is easy enough when grabbing a process, but when
rd_new() is called for a newly started process, or when an exec() happens, extra
complexity is introduced.  In a dynamically linked program, DT_DEBUG is
populated at dynamic linker initialization time, which has not yet happened; in
a statically linked program, _r_debug is there but r_brk has no address yet.  We
can handle the dynamic case by dropping a breakpoint on the entry point of the
program (as given by the AT_ENTRY auxiliary vector entry) and looking up
_r_debug et al then: this is called after the dynamic linker has initialized.
For statically linked programs, this is insufficient: glibc initializes the
dynamic linker in a statically linked program only when dlopen() is called.  So
every public rtld_db entry point gets the r_brk address via an r_brk() helper,
and returns RD_NOMAPS if it is not yet set up.  (At a later date, we may arrange
to automatically detect this case and start providing rtld activity reports as
soon as possible, but for now we simply return RD_NOMAPS until things are up.
So activity reporting is not very useful in statically linked programs yet,
though it does technically work.)

Once we have the necessary addresses, we can stipulate that in critical regions
of the program the link maps of a given traced child are guaranteed not to
change by dropping a breakpoint on the r_brk address and having the breakpoint
handler halt execution only if the consistency state changes to something other
than RD_CONSISTENT; on entry to such regions we pause until the state is
RD_CONSISTENT, and then can be certain that it cannot change until we resume the
process at the end of the critical region.  We carry out all link map traversal
inside such a region.  (We can also, naturally, use the same breakpoint to
implement dlactivity callbacks.)

The new ELF lookups needed by this machinery are mostly done by code in
libproc/elfish.c: where libelf machinery that varies depending on the 64-bitness
of the ELF file is needed (r_debug lookups), we thunk to code in
libproc/elfish_64.c or libproc/elfish_32.c accordingly.