Fix name/address lookup within breakpoint handlers, and stopping handler bugs.
If a breakpoint handler did a name or address lookup (via Pxlookup_by_name() or
Plookup_by_addr(), or otherwise did anything triggering a map rebuild and ELF
file read, the process would be resumed despite the fact the breakpoint handler
was still running, before the breakpoint management machinery had poked the
breakpoint instruction back in, possibly leading to one or more later hits of
the same breakpoint being missed.

This is because address lookup does a Ptrace()/Puntrace() pair so that it can
ensure the process is being ptraced so as to do a ptrace(PTRACE_GETMAPFD, ...),
and these calls are not quite inverses.  While a nested Ptrace() returns the
previous process state to pass into a subsequent Puntrace(), a top-level one
ignores the passed state and unconditionally resumes.  My assumption -- that the
process would be running before a top-level Ptrace(), so should always be set
running on top-level Puntrace() -- is trivially false on closer examination. It
could be stopped at a signal, dead, or, as here, stopped at a breakpoint.  If
the state passed in top a top-level Puntrace() is not PS_RUN (or the 'we have no
idea' 0 value, which should really only be used when resetting things after an
exec()), we should *not* resume execution, but leave resumption to the inverse
of whatever process had halted execution before the Ptrace().

Further, if a breakpoint handler requests a tracestop halt, we don't want to
resume execution when the top-level Puntrace() happens, or a sequence like this:

set up breakpoint
Ptrace()
Pwait()
breakpoint handler returns PS_TRACESTOP
Puntrace()

goes wrong, since the code outside the Ptrace()/Puntrace() pair will naturally
assume that we will have been halted when the handler fires, but it has, and we
were resumed!

To solve this, we track whether the breakpoint handler has returned PS_TRACESTOP
yet Pbkpt_continue() has not been issued.  This is a change in semantics, because
if a breakpoint was set up inside a Ptrace() / Puntrace() pair, you can no longer
rely on Puntrace() resuming you -- but this is easily solved: just add

if (Pbkpt_addr(P) != 0)
	Pbkpt_continue(P);

and you are home free.