Fix various problems in name lookup and map construction.
Predominant among this is a problem with the management of the ld.so and
executable mappings, searched via the PR_OBJ_LDSO and PR_OBJ_EXEC flags to
Pxlookup_by_name().  These were represented as map_info_t pointers, but
unfortunately these are pointers into an array, and were set during
Pupdate_maps() time, while the array's growth was still underway.  So they
tended to become pointers to garbage almost at once.  Instead, track indexes
into the mapping array (using -1 instead of NULL to indicate none), adjust all
callers accordingly, and add a Paddr2idx() for those places that want to convert
an address into a suitable index.  (We also reimplement the existing
Paddr2mptr() in terms of Paddr2idx(), so this is almost no new code.)

Another problem is an unexpected recursion.  The first name lookup in a
process (and some but perhaps not all later ones) triggers a Pupdate_maps(),
which includes an iteration over the link maps via rd_loadobj_iter() to figure
out the soname and lmid of every shared library.  rd_loadobj_iter() internally
calls rd_ldso_consistent_begin() to ensure the consistency of the load maps
across the iteration.  Unfortunately an upcoming commit changes this function to
do a name lookup for _rtld_global inside the dynamic linker... which itself
calls Pupdate_maps() and ends up calling rd_loadobj_iter() recursively.  While
recursive rd_loadobj_iter() calls have been tested and do work, we don't expect
to be called back from inside rd_ldso_consistent_begin() itself!  We end up
violating an invariant and getting stuck on a breakpoint forever: not good.  To
fix this, we have to decouple map iteration from map updating, so that we can
avoid doing a map update if we were called to look up names inside ld.so (since
we can be sure that such lookups don't care about lmids, or sonames, or anything
we can extract from rtld_db).  This is rather annoying to implement: not only do
we have to introduce a new lmids_valid field in the ps_prochandle, and a new
function Pupdate_lmids() to update the lmids, we also have to update every
single place that calls Pupdate_maps() to call Pupdate_lmids() too, excepting
only the one place where a PR_OBJ_LDSO might be done.

Also a couple of smaller bugfixes, handling failed realloc()s in Pupdate_map()
and avoiding spurious errors when we encounter text mappings that aren't ELF
files.