Quash a potential use of uninitialized data in symbol lookup.
If symtab_getsym() failed to find a symbol with a given index, sym_by_addr()
would use uninitialized data.  This can only happen if the ELF file is malformed
or if there are bugs in symbol table reading, but still it is worth fixing.

Also quash a couple of other incorrect warnings.