Fix unintended breakpoint handler / name lookup interaction.
Pbuild_file_symtab(), called to read in ELF data before doing symbol or name
lookups, has to call Ptrace()/Puntrace() in order to halt the process around a
PTRACE_GETMAPFD on each mapping (since PTRACE_GETMAPFD, like all informational
ptrace() requests, requires the tracee to be in ptrace-stop state).  Ptrace()
and Puntrace() can both issue a Pwait() to update libproc's idea of process
state, and this can execute breakpoint handlers.  Breakpoint handlers, in their
turn, can look up names and addresses -- which will recurse back into
Pbuild_file_symtab()!  This is not an infinite recursion, because once the
process is stopped on a breakpoint handler, it is stopped, and no more
breakpoint handlers will run -- but it can lead to a false failure to look up
some names in the breakpoint handler, because the file structure that the outer
call to Pbuild_file_symtab() is working over claims to be initialized, but is
not.

The fix is to temporarily mark the file as not initialized around the calls to
Ptrace() and Puntrace() in Pbuild_file_symtab(), and check after those calls to
see if it has been quietly marked as initialized by a recursive call.  Then
everything, even breakpoint handlers, can look up names reliably.