Fix problems removing breakpoints while Ptrace()-stopped on them.
This is a followup to f361f3ec60cb4cdac0eb8d9292a47a2e91503dc0.

If you call Ptrace(P, 1) the process blocks.  It is supposed to remain blocked
until you explicitly unblock it (via Pbkpt_continue(), SIGCONT, PTRACE_CONT
or the like) or until the corresponding Puntrace() call (assuming you pass in
the state returned from the corresponding Ptrace() call correctly).

This was working in all but one case.  Removal of a breakpoint with Punbkpt() is
supposed to resume execution if it was stopped at that point by a breakpoint
handler that returned PS_TRACESTOP.  This was working -- but it was also
reverting the effect of any Ptrace() too!

The effect was that this sequence, inside a breakpoint handler on address B:

    s = Ptrace(P, 1);
    Punbkpt(P, B);
    /* blah */
    Puntrace(P, s);

is resumed across 'blah', even though you've just done a blocking Ptrace() and
execution should be suspended until the corresponding Puntrace().  This was
breaking rtld_db's process-execution trapping code, which does just that.

Note that the sequence

    s = Ptrace(P, 0);
    Punbkpt(P, B);
    /* blah */
    Puntrace(P, s);

*should* resume across 'blah'.  However,

    s1 = Ptrace(P, 1);
    s2 = Ptrace(P, 0);
    Punbkpt(P, B);
    /* blah */
    Puntrace(P, s2);
    Puntrace(P, s1);

should not, because a blocking Ptrace() is in effect (the outermost one).

To fix this, we do something similar to what we did with 'bkpt_halted' in the
earlier commit, introducing a new bit of state 'ptrace_halted', setting it when
a halting Ptrace() is invoked, and clearing it when we are resumed by Puntrace().
Now Punbkpt() can just check that to see if it ought to resume when removing a
breakpoint we are halted on.