Properly drain SIGSTOPs over exec().
When a blocking Ptrace() call is made, and the process is not already stopped,
it stops the process by sending it a SIGSTOP, and then doing a blocking Pwait()
for it.  Unfortunately this can race with the child doing an exec(), hitting the
process after exec() and halting it, even though our own exec-detection code
detaches the Ptrace() automatically upon child exec().  So, maintain a count of
all SIGSTOPs sent but not yet consumed by Pwait(), and another count of SIGSTOPs
sent before exec() but not received until after it (copied from the first count
when an exec() is detected).  Then, in SIGSTOP handling, decrease the pending
SIGSTOP count whenever it is nonzero, and whenever the sigstop-pre-exec count is
nonzero, automatically continue the process to reverse their unwanted effect.

(A side benefit of this is that we can now distinguish our own SIGSTOPs from
third-party SIGSTOPs, so if you SIGSTOP a process we are tracing we do not get
confused and think that we placed it in blocking trace stop state.)