Do not SIGSTOP already SIGSTOPped processes when Ptrace()ing them.
When Ptrace()ing a process, Ptrace() tries to get it into a consistent
state (stopped, or not stopped, as requested), whether or not the process is
already under the control of ptrace()-the-syscall.  If the process is
already being ptrace()d, this means it has to hit it with a SIGSTOP to
stop it.  This has already been the source of multiple bugs regarding
SIGSTOPs crossing exec() and the like, but now we have another one.

We already avoid sending any kind of signal if the process is already in
trace-stop state, but we should also avoid sending it if the process is in stop
state, since sending a SIGSTOP to something that is already SIGSTOPped is
pointless, and since we follow that up with a blocking wait for the process's
state to stabilize, and if it is already SIGSTOPped that wait will never return.