Symbol search scope support.
When looking up symbols (whether by dlsym(), or triggered by a use of & in a
program), the dynamic linker searches a list of *scopes* (each of which
corresponds to a link map).  Every shared library can have its own scope, though
most don't: shared libraries loaded new lmids by dlmopen() routinely do, as do
those linked with -Bsymbolic.

For DTrace this is mostly irrelevant for now: foo`bar shouldn't be following a
scope in any case.  But for global lookups, we want to respect scopes to some
degree, so we need access to the dynamic linker's scopes.  This is somewhat
annoying, because they're in the private part of the link map, and before now we
have not needed access to anything but its public part.  Worse yet, they are
buried after the l_info pointers, which is an array sized by the ELF dynamic
tags (including the platform-specific one, which is not public in <elf.h> so we
have to add a definition in <platform.h> for it, even though it is zero on all
platforms supported by DTrace at the moment).  So if glibc starts to support a
new ELF dynamic tag in the future, the offsets we will need will change.  This
happens every few years: so not very often, but nonetheless we must consider
fallback behaviour in case the lookups fail.  Still, this is much better than
encoding knowledge of DT_SYMBOLIC, DT_GROUP, multiple lmids and whatever other
strange exceptional symbol lookup rules might be added to ELF in the future: far
better to just rely on the dynamic linker to do it, even if it means we have
to cope with changing offsets now and then.

Looking at the API for this before we look at the implementation, we have three
new fields in rd_loadobj_t, the type returned from the rd_loadobj_iter()
iteration function:

 - rl_scope, an array of pointers into the victim's address space representing
   link maps which should be searched for symbols, in order, if the symbol is
   not found in the current link map; to avoid calling the allocator over and
   over, this is repeatedly overwritten, and realloc()ated only when it needs
   to expand, not shrink
 - rl_nscopes, the numer of scopes in rl_scope which are currently valid
 - rl_default_scope, true if this scope is the same as the 'default scope' for
   this lmid (see below)
 - rl_nscopes_alloced, not meant for external use, tracking the largest number
   of scopes that have ever been in rl_nscopes, so we know when we have to
   realloc() it

rd_get_scope(), which indexes into the nscopes array for you and turns the Nth
member into another rd_loadobj_t object, just like you would be passed from
rd_loadobj_iter().  You have to pass in an all-NULL one on the first call, and
free() its rl_scope member after you're done with all calls.

Internally, rd_get_scope() is just a thin wrapper around rd_get_link_map() and
rd_get_loadobj_link_map(), two functions refactored out of rd_loadobj_iter().
The first turns an address in a child into a link map, and the second turns this
into a rd_loadobj_t object, looking up victim addresses and reading in the
scope elements from the link map, doing all as-needed realloc()ing.

rd_loadobj_iter() has another problem to deal with.  Link maps fairly often have
empty search scopes, but that doesn't mean you don't look anywhere for symbols
if this search fails: instead, it means that you should look at the first search
scope in this lmid and search that instead: a default. This means that glibc
doesn't need to waste memory allocating identical link maps over and over for
ordinary shared libraries: instead it just allocates one and looks in it as
needed. So, as we walk over the lmids, we have to remember the first search
scope for every lmid, and when needed temporarily copy it back into the
rd_loadobj_t structure we pass to our callback.