Fix link line for programs in test/utils.
The programs in this directory call libdtrace.so published APIs only.  libdtrace
already names all its own dependencies in DT_NEEDED entries, so we don't need to
name them again.  (Naming libproc and libport is particularly risky, since these
are not shared libraries but build libraries only, so we were using the wrong
names for them, and ended up linking against the libproc from procps, if that
was installed, or failing otherwise!)