Do not run libproc tests in check-installed mode by default.
These tests are unusual in that they depend on build libraries explicitly.  In a
clean source tree, no such libraries exist, so a `make check-install' in a clean
source tree was going off and building all of DTrace, rather than (as intended)
just building triggers.

This isn't really good enough, and was not intended.  So split up the triggers
into `external' and 'internal' triggers, the former of which are run during
`make check-installed', the latter of which depend on internal implementation
details, and are not.  There is a new `external-triggers' Makefile rule to build
the external triggers only, which is run by the check-installed rule and all its
related rules.

We also move the libproc tests into a new 'internals' testsuite, and run them by
default, so that the -installed- rules can stop running them.

(None of this prevents you from running the internals tests in check-installed
mode by hand, via ./runtest.sh --use-installed.  It's only the makefile rule that
skips them.)