Add rpath when building victims tested by libproc internal tests.
The victims analyzed by the libproc internal tests have to load shared
libraries, which means they have to be locatable.  In normal test runs, this is
easy, as LD_LIBRARY_PATH points through build/, but when explicitly running the
libproc tests in -installed mode, this is not true, and the testcases fill up
the log with millions of lines of cannot-find-shared-library errors.  To fix
this, add an rpath to the victims, so they always look in build/ for their
shared libraries even if the parent process does not.