Correctly diagnose negative option values.
Some DTrace options are prohibited from being negative.  The test for this broke
recently because it was relying on getting a negative number (cast to unsigned
long long) back from strtoull() when a negative number was passed in.  This is
not guaranteed: the result you get in this case is implementation-dependent and
varies between classic Unixes and Solaris and glibc, and indeed between glibc
versions.

So call strtoll() first, just to check negativity.