Revamp and split up DTrace headers.  Add ioctl() debugging machinery.
This is the userspace side of the header splitting work.  As with the kernel
side, it removes virtually all content from the userspace copy of dtrace.h in
uts/common/sys, replacing it with an #include of the shared headers provided by
the kernel module.  dtrace_types.h is updated to remove a couple of dead types
and to add the definition of dtrace_id_t, which is still not shared with userspace.
dtrace_ioctl.h is gone, provided entirely by the kernel side.  One definition
(of dtrace_probespec_t) which had crept into the userspace copy of dtrace.h but
has no relevance to the kernel side at all was moved into libdtrace/dtrace.h.
All this is installed in /usr/include/sys as before, to avoid disrupting existing
userspace.

As mentioned in the kernel-side commit log, we also install a trivial copy of
dtrace_cpu_defines.h into linux/, for the sake of the shared
linux/dtrace/ioctl.h header.

The ioctl()-debugging machinery is now triggered on every dtrace startup and
printed when DTRACE_DEBUG is turned on.  This should make it easier to compare
ioctl() type sizes between kernel and userspace in future.

There's a bit of gardening of the specfile in here, too, removing unnecessary
%defattr mode forcing, adding a Group to the development package, and boosting
the version number.