Add breakpoint notifiers.
These are just like breakpoints, except that they return void, so cannot affect
control flow, and there can be more than one of them at a given address. They
can have cleanups, just like breakpoints, but there is no guarantee that the
after_singlestep property is respected: any breakpoint on that address will
supersede it.  (This could be fixed, if necessary, but right now it would
add complexity we don't use.)

Implementation is fairly trivial: the largest tricky part is that now it is
possible to have a 'breakpoint' with no handler, only notifiers.  Also, calling
the cleanup functions has now got verbose enough that I've factored it out into
a separate function.