Do not crash on PR_OBJ_EVERY name lookup if rtld_db cannot be initialized.
Pupdate_symsearch() was using the rtld_db map-iteration pointer (fptr->file_lo)
without checking if it was NULL, which it could legitimately be if it was too
early in process execution to initialize rtld_db.  This caused crashes if you
did a symbol lookup with PR_OBJ_EVERY that early on.  (Such a symbol lookup
could not follow symbol scopes, but should still be able to do a linear scan.
It should certainly not crash.)