A couple of symbol search scope bugfixes.
Firstly, an unfortunate typo led to the ld.so-scoped part of the symbol search
often (but not always) being skipped: it fell through to the linear search
Solaris always used to do, so most things still worked.  Another unfortunate
typo, spotted by valgrind, led to an off-by-one bug in the ld.so-scoped symbol
search code and could easily have led to crashes (once the first bug was fixed).