ustack() test improvements.
When we get ustack() working, it will have a few restrictions. Among other
things, it can't work if the process is dead, and it can't work if the process
doesn't have a symbol table or if its functions have all been merged by
sibcalling and inlining or because they are all static and call each other only
once.

So turn all those optimizations off, make them non-static, and stop the
ustack test exit()ing right after the ustack().