Fix Ptrace()/Puntrace() imbalance.
It is crucial that every Ptrace() call be accompanied by exactly one Puntrace().

Pbuild_file_symtab() was failing to do this in one case, if its first Ptrace()
led to a breakpoint hit and a recursive call to Pbuild_file_symtab() such that
the outermost call to Pbuild_file_symtab() didn't need to do anything any more.