Prohibit self-grabs.
If a ustack() or similar operation is done from a BEGIN probe, the effective
process is the DTrace process itself.  But you can't ptrace() yourself!  In
theory we could reimplement (almost) everything we've done with ptrace() using
direct memory access, but doing that for this one case is pointless.

It's easier just to ban self-grabs: DTrace then falls back to reporting pure
addresses in ustack() et al.  (Once we have noninvasive grabs, we'll be able to
improve on this a bit.)