Ensure multiple passes through dtrace -G work.
Due to how relocations for USDT probe locations were handled, doing a second
pass through 'dtrace -G' would cause probes to no longer be recognized.  This
commit ensures that we recognize relocations of type NONE as probe locations
that were already processed rather than ignoring them.

Also add an explicit define for _DTRACE__VERSION in the generated header file
from 'dtrace -h' invocations.  This seems silly, but it is there to give people
a skeleton to work with if they want to pre-generate header files and select
whether to actually compile on the probes at a later time.