Fix ustack() of newly-started processes.
Newly-started processes need to update their list of loaded mappings as soon as
the dynamic linker has initialized.  We have a breakpoint dropped at this point,
but were not initializing the dynamic linker then because it triggered a
deadlock inside libproc.  With this deadlock now fixed, we can call it, and pull
the DT_NEEDED libraries in.

Also fix a couple of silly (harmless) format string bugs in debugging
statements, and drop some very noisy and no-longer-necessary debugging.