Set priority on DRTI constructor, report SKIPs, add fasttrap, add showUSDT.
Updated compiler.h to specify a priority (0) for the constructor used in DRTI.

Updated runtest.sh to properly report SKIPs in the summary file.

Added fasttrap to the list of modules to load for the testsuite runs.

Added the showUSDT tool to dump the content of the DOF elf section in objects
(usually executables) that contains USDT probes.