Temporarily disable use of dtrace_ioctl_sizes().
Due to the need to remove the function definition for dtrace_ioctl_sizes() from
ioctl.h to ensure that the header file (and dtrace.h) can be used in other
code, this call had to be disabled.  If/when a new location for the function
is provided, this can be re-enabled.