ISA-dependent dispatch machinery for libproc, and first-argument fetcher.
We already have machinery in bliproc that conditionalizes based purely on ELF
class (32- or 64-bit), building both classes from a common source file with
preprocessor defines (elfish*.[ch]). However, we do not have machinery to allow
for code that differs completely depending on the ELF class or ISA (e_machine).
Such machinery cannot be preprocessor-built: in some cases (but perhaps not all)
multiple instances must be compiled in simultaneously, and selected at runtime
(32/64-bit is one example, as might be big/little-endian on SPARC: but SPARC ISA-
dependent code may well not compile on x86 at all).

So, introduce isadep.c, which has machinery to dispatch to underlying functions
according to e_machine and bitness: with a very few changes (to be done when
necessary) we can arrange to package underlying functions that cannot be called
on specific machines into source files that are not compiled on such machines,
and use weak symbols to avoid introducing any machine-dependent #ifdefs into
isadep.c at all.

But for now it is much simpler, containing one single public function,
Pread_first_arg(), to be used by later commits in this series. This function can
be called when execution is stopped at the start of a (non-inlined) function,
and returns the value of its first argument, which must be a pointer, in an
ISA-dependent fashion.