Avoid trying to initialize rtld_db on a dead process.
We attempt to initialize rtld_db on every run through the main process control
loop, just in case we couldn't initialize it earlier because the process hadn't
got far enough to initialize the dynamic linker (and getting it that far might
take several breakpoint traps, which requires Pwait() calls, thus trips through
said control loop).

However, if the process dies before getting far enough to initialize rtld_db --
fairly easy if it's statically linked and never dlopen()s anything, in which
case the dynamic linker is never initialized -- we try to initialize rtld_db
*before* checking if we just got a notification because the process just died.
This is harmless, but pointless, and until recently emitted ugly debug spam on
stderr.

(There is still a narrow race where we can try to initialize rtld_db on a dead
process: if it is a statically linked process with an uninitialized dynamic
linker that gets hit with a signal, say, and the signal causes the process to
die, but the system is busy enough that DTrace does not get back round the
control loop into the waitfd poll() before the death happens. This is harmless,
but worth nothing anyway.)