Improved evaltime=main support.
Before now, evaltime=main has been restricted to working on executables with
dynamic symbol tables.  Since very few executables have dynamic symbol tables,
this makes the option less than useful.

This is as good as we can do for statically linked programs, but the vast
majority of programs are dynamically linked.  In dynamically linked programs,
the statically linked startup code calls __libc_start_main(), which is in
libc.so and thus always has a dynamic symbol table entry we can find.  The first
argument to this function is the very address we need, the address of main().
Getting this argument is ISA-dependent (see earlier commits in this series), but
not particularly onerous, and completely baked into the ABI and unchangeable by
virtue of being statically linked into countless programs.  So we can definitely
depend on it.

So now, evaltime=main does the following (with associated dt_proc_attach()
states):

 - halt execution at process invocation
 - ATTACH_START: drop a breakpoint at AT_ENTRY, and resume, later hitting:
 - dt_break_prepare_drop_main(): ATTACH_ENTRY: drop a breakpoint on
   __libc_start_main() if dynamically linked, and resume, later hitting:
 - dt_break_libc_start_main(): ATTACH_FIRST_ARG_MAIN: drop a breakpoint
   on the address given in the first argument of __libc_start_main().

If any of these lookups fail, or the program is statically linked, we drop
back to the old machinery of dropping a breakpoint on main() if we can look
it up (identified by the state ATTACH_DIRECT_MAIN).  Finally, we fall back
to preinit, but this should now be a very rare failure mode, seen only on
stripped statically linked binaries.

We have new tests for all of this (the 32-bit test is hived off and XFAILed
until such time as syscall tracing works for the 32-bit compat syscall table).