Set the default value of the evaltime option to postinit.
USDT depends on constructor execution, which must happen before DTrace
rendezvouses and starts tracing.  In Solaris, the default evaltime of preinit
suffices for this, becuase preinit on Solaris runs after constructors, but
before #pragma init. On Linux, preinit runs *before* constructors (but after
dynamic linker initialization).  So we must set the default to postinit or main,
which is now practical since these should now work fairly reliably for the
majority of processes.