Fix testsuite option extraction.
Option extraction was failing to strip trailing spaces from option values (which
can be quite common if they are specified in a comment inside a .d file) and
also gave hilariously bad results if an option's name was a subset of another's
(an option like that will be added shortly).

The sequence of regexps we now grep for should catch
/* @@foo */
/* @@foo: bar */
/* @@foo :bar */
# @@foo (in all variants)

without catching

/* @@foo-bar: */

or anything like it.