Fix debugging whine about processes in trace stop.
When libproc (in the master thread) issues a SIGSTOP to a child process and then
Pwait()s for it, it is possible that the control thread will awaken with both
the proxy request and the state change outstanding in its poll request
structure.  In this case, the Pwait() will do nothing, but we will then fall
into a case statement which issues a debugging whine about the
PS_TRACESTOP (because if something else had placed the process into tracestop,
we would be deadlocked).

This is inappropriate in this case and should be avoided.